Cleanup: Make _next_round_robin uint64_t

+++ iocore/eventsystem/I_EventProcessor.h
@@ -303,9 +303,9 @@ public:
   /// The thread group ID is the index into an array of these and so is not stored explicitly.
   struct ThreadGroupDescriptor {
     std::string _name;                               ///< Name for the thread group.
-    int _count                = 0;                   ///< # of threads of this type.
-    std::atomic<int> _started = 0;                   ///< # of started threads of this type.
-    int _next_round_robin     = 0;                   ///< Index of thread to use for events assigned to this group.
+    int _count                 = 0;                  ///< # of threads of this type.
+    std::atomic<int> _started  = 0;                  ///< # of started threads of this type.
+    uint64_t _next_round_robin = 0;                  ///< Index of thread to use for events assigned to this group.
     Que(Event, link) _spawnQueue;                    ///< Events to dispatch when thread is spawned.
     EThread *_thread[MAX_THREADS_IN_EACH_TYPE] = {}; ///< The actual threads in this group.
     std::function<void()> _afterStartCallback  = nullptr;
